results/rebuttal/run-20231111-034650_LLMatic-zero_shot_donnemartin-system-design-oop-bookclub_0_20-runs_log/bookclub/9/LLMatic/app.py
from flask import Flask, request

app = Flask(__name__)

# Mock database
DATABASE = {
	'users': {},
	'book_clubs': {},
	'meetings': {},
	'discussions': {},
	'book_selections': {},
	'user_profiles': {},
	'recommendations': {},
	'admin_actions': {},
	'notifications': {},
	'resources': {}
}

# User class
class User:
	def __init__(self, username, password):
		self.username = username
		self.password = password
		self.following = []

	def authenticate(self, password):
		return self.password == password

	def update_profile(self, password):
		self.password = password

	def follow_user(self, user):
		self.following.append(user)

	def get_info(self):
		return {'username': self.username, 'password': self.password, 'following': [user.username for user in self.following]}

	def send_notification(self, notification):
		# Mock sending of notification
		return 'Notification sent'

	def customize_notifications(self, settings):
		# Mock customization of notifications
		return 'Notifications customized'

# Recommendation Engine
class RecommendationEngine:
	def __init__(self):
		self.recommendations = {}

	def generate_recommendations(self, user):
		# Mock recommendation generation based on user's reading history and popular books
		self.recommendations[user.username] = ['Book1', 'Book2', 'Book3']

	def get_recommendations(self, user):
		return self.recommendations.get(user.username, [])

# Initialize Recommendation Engine
RECOMMENDATION_ENGINE = RecommendationEngine()

# Admin class
class Admin:
	def __init__(self, username, password):
		self.username = username
		self.password = password

	def manage_book_clubs(self, action, book_club):
		# Mock management of book clubs
		return 'Book club managed successfully'

	def manage_user_accounts(self, action, user):
		# Mock management of user accounts
		return 'User account managed successfully'

	def remove_inappropriate_content(self, content):
		# Mock removal of inappropriate content
		return 'Inappropriate content removed successfully'

	def get_analytics(self):
		# Mock retrieval of analytics
		return 'Analytics retrieved successfully'

# Remaining classes and routes...

@app.route('/view_profile', methods=['GET'])
def view_profile():
	username = request.args.get('username')
	if username in DATABASE['users']:
		return DATABASE['users'][username].get_info()
	else:
		return 'User not found', 404

@app.route('/update_profile', methods=['POST'])
def update_profile():
	data = request.get_json()
	username = data['username']
	password = data['password']
	if username in DATABASE['users']:
		DATABASE['users'][username].update_profile(password)
		return 'Profile updated successfully', 200
	else:
		return 'User not found', 404

@app.route('/follow_user', methods=['POST'])
def follow_user():
	data = request.get_json()
	username = data['username']
	user_to_follow = data['user_to_follow']
	if username in DATABASE['users'] and user_to_follow in DATABASE['users']:
		DATABASE['users'][username].follow_user(DATABASE['users'][user_to_follow])
		return 'User followed successfully', 200
	else:
		return 'User not found', 404

@app.route('/generate_recommendations', methods=['POST'])
def generate_recommendations():
	data = request.get_json()
	username = data['username']
	if username in DATABASE['users']:
		RECOMMENDATION_ENGINE.generate_recommendations(DATABASE['users'][username])
		return 'Recommendations generated successfully', 200
	else:
		return 'User not found', 404

@app.route('/view_recommendations', methods=['GET'])
def view_recommendations():
	username = request.args.get('username')
	if username in DATABASE['users']:
		return RECOMMENDATION_ENGINE.get_recommendations(DATABASE['users'][username])
	else:
		return 'User not found', 404

# Admin routes

@app.route('/manage_book_clubs', methods=['POST'])
def manage_book_clubs():
	data = request.get_json()
	username = data['username']
	action = data['action']
	book_club = data['book_club']
	if username in DATABASE['users'] and isinstance(DATABASE['users'][username], Admin):
		return DATABASE['users'][username].manage_book_clubs(action, book_club)
	else:
		return 'User not found or not an admin', 404

@app.route('/manage_user_accounts', methods=['POST'])
def manage_user_accounts():
	data = request.get_json()
	username = data['username']
	action = data['action']
	user = data['user']
	if username in DATABASE['users'] and isinstance(DATABASE['users'][username], Admin):
		return DATABASE['users'][username].manage_user_accounts(action, user)
	else:
		return 'User not found or not an admin', 404

@app.route('/remove_inappropriate_content', methods=['POST'])
def remove_inappropriate_content():
	data = request.get_json()
	username = data['username']
	content = data['content']
	if username in DATABASE['users'] and isinstance(DATABASE['users'][username], Admin):
		return DATABASE['users'][username].remove_inappropriate_content(content)
	else:
		return 'User not found or not an admin', 404

@app.route('/get_analytics', methods=['GET'])
def get_analytics():
	username = request.args.get('username')
	if username in DATABASE['users'] and isinstance(DATABASE['users'][username], Admin):
		return DATABASE['users'][username].get_analytics()
	else:
		return 'User not found or not an admin', 404

# Notification routes

@app.route('/send_notification', methods=['POST'])
def send_notification():
	data = request.get_json()
	username = data['username']
	notification = data['notification']
	if username in DATABASE['users']:
		return DATABASE['users'][username].send_notification(notification)
	else:
		return 'User not found', 404

@app.route('/customize_notifications', methods=['POST'])
def customize_notifications():
	data = request.get_json()
	username = data['username']
	settings = data['settings']
	if username in DATABASE['users']:
		return DATABASE['users'][username].customize_notifications(settings)
	else:
		return 'User not found', 404

if __name__ == '__main__':
	app.run(debug=True)
